+++ tests/files/src/0488_crash.php
@@ -1,0 +1,4 @@
+<?php
+function test488($definitions) {
+    $groupDef = ['prefix'] = $definitions->getPrefix();
+}